rozdzielenie pliku views
