+++ main.py
@@ -8,7 +8,9 @@ from flask import Flask
 # importy nasze
 from views import pages
 from views_admin import pages_admin
+from views_ajax import pages_ajax
 from views_login_system import pages_log_sys, login_manager
+from views_user import pages_user
 from mail_sending import mail
 from db_models import db
 from config import DB, APP, EMAIL
@@ -27,7 +29,9 @@ app.config['MAIL_DEFAULT_SENDER'] = EMAIL.DEFAULT_SENDER
 
 app.register_blueprint(pages)
 app.register_blueprint(pages_admin)
+app.register_blueprint(pages_ajax)
 app.register_blueprint(pages_log_sys)
+app.register_blueprint(pages_user)
 login_manager.init_app(app)
 mail.init_app(app)
 db.init_app(app)
