+++ templates/navbar.html
@@ -13,7 +13,7 @@
         <a class="nav_right" href="{{url_for('pages_log_sys.logout')}}">Wyloguj</a>
     </div>
     <div class="nav_button_right" style="min-width: 100px">
-        <a class="nav_log_in_as" href="{{ url_for('pages.user_page') }}">
+        <a class="nav_log_in_as" href="{{ url_for('pages_user.user_page') }}">
             Zalogowany jako:<br>
             {% if current_user.name and current_user.last_name %}
                 {{ current_user.name }} {{ current_user.last_name }}
